airhub-ml/requirements.txt
fastapi==0.104.1
torch>=2.0.0
uvicorn==0.23.2
numpy<=1.24.3
scikit-learn>=1.7.2
pandas==2.0.3
tensorflow<=2.12.0
flwr==1.5.0  # Flower for federated learning
requests==2.31.0
python-dotenv==1.0.0
pydantic==2.4.2
matplotlib==3.7.2
pytest==7.4.0